Use MobX for model\app.dart and remove MainWindowCubit
