+++ lib/widgets/main_window/main_window.dart
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2021 - 2022 Joshua Wade
+  Copyright (C) 2021 - 2023 Joshua Wade
 
   This file is part of Anthem.
 
@@ -19,16 +19,19 @@
 
 // cspell:ignore autofocus
 
-import 'package:anthem/helpers/id.dart';
-import 'package:anthem/widgets/basic/overlay/screen_overlay.dart';
-import 'package:anthem/widgets/basic/overlay/screen_overlay_cubit.dart';
-import 'package:anthem/widgets/main_window/tab_content_switcher.dart';
-import 'package:anthem/widgets/basic/menu/menu.dart';
-import 'package:anthem/widgets/main_window/window_header.dart';
-import 'package:flutter/widgets.dart';
+import 'package:anthem/model/store.dart';
+import 'package:anthem/widgets/main_window/main_window_controller.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:anthem/widgets/main_window/main_window_cubit.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
+
+import 'package:anthem/helpers/id.dart';
+import 'package:anthem/widgets/basic/menu/menu.dart';
+import 'package:anthem/widgets/basic/overlay/screen_overlay_cubit.dart';
+import 'package:anthem/widgets/basic/overlay/screen_overlay.dart';
+import 'package:anthem/widgets/main_window/tab_content_switcher.dart';
+import 'package:anthem/widgets/main_window/window_header.dart';
 
 class MainWindow extends StatefulWidget {
   const MainWindow({Key? key}) : super(key: key);
@@ -40,12 +43,15 @@ class MainWindow extends StatefulWidget {
 class _MainWindowState extends State<MainWindow> {
   bool isTestMenuOpen = false;
   MenuController menuController = MenuController();
+  MainWindowController controller = MainWindowController();
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<MainWindowCubit, MainWindowState>(
-        builder: (context, state) {
-      return BlocProvider<ScreenOverlayCubit>(
+    final store = AnthemStore.instance;
+
+    return Provider.value(
+      value: controller,
+      child: BlocProvider<ScreenOverlayCubit>(
         create: (context) => ScreenOverlayCubit(),
         child: ScreenOverlay(
           child: RawKeyboardListener(
@@ -53,20 +59,20 @@ class _MainWindowState extends State<MainWindow> {
             autofocus: true,
             onKey: (e) {
               final type = e.runtimeType.toString();
-
+    
               final keyDown = type == 'RawKeyDownEvent';
               final keyUp = type == 'RawKeyUpEvent';
-
+    
               final ctrl = e.logicalKey.keyLabel == "Control Left" ||
                   e.logicalKey.keyLabel == "Control Right";
               final alt = e.logicalKey.keyLabel == "Alt Left" ||
                   e.logicalKey.keyLabel == "Alt Right";
               final shift = e.logicalKey.keyLabel == "Shift Left" ||
                   e.logicalKey.keyLabel == "Shift Right";
-
+    
               final keyboardModifiers =
                   Provider.of<KeyboardModifiers>(context, listen: false);
-
+    
               if (ctrl && keyDown) keyboardModifiers.setCtrl(true);
               if (ctrl && keyUp) keyboardModifiers.setCtrl(false);
               if (alt && keyDown) keyboardModifiers.setAlt(true);
@@ -78,31 +84,42 @@ class _MainWindowState extends State<MainWindow> {
               color: const Color(0xFF2A3237),
               child: Padding(
                 padding: const EdgeInsets.all(3),
-                child: Column(
-                  children: [
-                    WindowHeader(
-                      selectedTabID: state.selectedTabID,
-                      tabs: state.tabs,
-                      setActiveProject: (ID id) {
-                        context.read<MainWindowCubit>().switchTab(id);
-                      },
-                      closeProject: (ID id) {
-                        context.read<MainWindowCubit>().closeProject(id);
-                      },
-                    ),
-                    Expanded(
-                      child: TabContentSwitcher(
-                        tabs: state.tabs,
-                        selectedTabID: state.selectedTabID,
+                child: Observer(builder: (context) {
+                  final tabs = store.projectOrder
+                      .map<TabDef>(
+                        (projectID) => TabDef(
+                          id: projectID,
+                          title: store.projects[projectID]!.id,
+                        ),
+                      )
+                      .toList();
+    
+                  return Column(
+                    children: [
+                      WindowHeader(
+                        selectedTabID: store.activeProjectID,
+                        tabs: tabs,
+                        setActiveProject: (ID id) {
+                          store.setActiveProject(id);
+                        },
+                        closeProject: (ID id) {
+                          store.closeProject(id);
+                        },
                       ),
-                    ),
-                  ],
-                ),
+                      Expanded(
+                        child: TabContentSwitcher(
+                          tabs: tabs,
+                          selectedTabID: store.activeProjectID,
+                        ),
+                      ),
+                    ],
+                  );
+                }),
               ),
             ),
           ),
         ),
-      );
-    });
+      ),
+    );
   }
 }
